Issue #1404: And more corrections and improvements
