+++ api/src/main/java/com/epam/pipeline/acl/datastorage/FileShareMountApiService.java
@@ -30,12 +30,12 @@ public class FileShareMountApiService {
     private FileShareMountManager fileShareMountManager;
 
     @PreAuthorize(AclExpressions.ADMIN_ONLY)
-    public FileShareMount save(FileShareMount fileShareMount) {
+    public FileShareMount save(final FileShareMount fileShareMount) {
         return fileShareMountManager.save(fileShareMount);
     }
 
     @PreAuthorize(AclExpressions.ADMIN_ONLY)
-    public void delete(Long id) {
+    public void delete(final Long id) {
         fileShareMountManager.delete(id);
     }
